+++ LibOS/shim/src/fs/str/fs.c
@@ -90,7 +90,7 @@ int str_close (struct shim_handle * hdl)
     return 0;
 }
 
-int64_t str_read (struct shim_handle * hdl, void * buf, size_t count)
+ssize_t str_read (struct shim_handle * hdl, void * buf, size_t count)
 {
     int64_t ret = 0;
 
@@ -115,8 +115,8 @@ int64_t str_read (struct shim_handle * hdl, void * buf, size_t count)
     if (!strhdl->ptr)
         strhdl->ptr = data->str;
 
-    uint64_t offset = strhdl->ptr - data->str;
-    uint64_t remain = data->len - offset;
+    off_t  offset = strhdl->ptr - data->str;
+    size_t remain = data->len - offset;
 
     if (count >= remain) {
         memcpy(buf, strhdl->ptr, remain);
@@ -135,7 +135,7 @@ out:
     return ret;
 }
 
-int64_t str_write (struct shim_handle * hdl, const void * buf, size_t count)
+ssize_t str_write (struct shim_handle * hdl, const void * buf, size_t count)
 {
     if (!(hdl->acc_mode && MAY_WRITE))
         return -EACCES;
@@ -184,8 +184,7 @@ int64_t str_write (struct shim_handle * hdl, const void * buf, size_t count)
     return count;
 }
 
-int str_seek (struct shim_handle * hdl, off_t offset,
-              int whence)
+off_t str_seek (struct shim_handle * hdl, off_t offset, int whence)
 {
     struct shim_str_handle * strhdl = &hdl->info.str;
